refactor(ui): move RagDetailedStatus into rag-types so the presenter is standalone
Addresses CodeRabbit review feedback on #1147: the shared presenter
(src/ui/components/rag-status-panel.ts) imported its core RagDetailedStatus
type from one of its own consumers (rag-status-modal.ts), while that consumer
imports the presenter's render functions — a bidirectional module dependency
mitigated only by import-type erasure.

RagDetailedStatus is pure domain data (it already only references RagIndexStatus
and FailedFileEntry, both defined in src/services/rag-types.ts), so move it
there. The presenter and both modal consumers now depend on the domain type
rather than on each other. Type-only change: main.js remains byte-identical.
