+++ src/ui/background-tasks-modal.ts
@@ -1,7 +1,7 @@
 import { App, Modal, Notice, setIcon } from 'obsidian';
 import type ObsidianGemini from '../main';
 import type { BackgroundTask } from '../services/background-task-manager';
-import type { RagDetailedStatus } from './rag-status-modal';
+import type { RagDetailedStatus } from '../services/rag-types';
 import type { ProgressListener } from '../services/rag-types';
 import type { RagIndexingService } from '../services/rag-indexing';
 import { getErrorMessage } from '../utils/error-utils';
